+++ install/mediabot.sql
@@ -825,7 +825,8 @@ INSERT INTO `CHANSET_LIST` (`id_chanset_list`, `chanset`) VALUES
 (23, 'LangES'),
 (24, 'Wit'),
 (25, 'Spark'),
-(26, 'VDM');
+(26, 'VDM'),
+(27, 'DansTonChat');
 
 --
 -- PUBLIC_COMMANDS_CATEGORY — default categories